Updates to logging setup in utilities
This commit fixes a bug where program loggers created by get_logger would cause the root logger to always be set up in a similar manner, which would mean that all library logging was also turned on all the time.

Due to the change above, there was also added convinience functions for activating library logging in a more controlled manner.