Several updates to apps/tof/tof.py
Communication with pull sockets has been replaced with the
DateDataPullClient module for simplicity and stability. The tof
voltages via the Agilent unit are now acquired directly from the
unit instead of via rasppi74.
The measurements acquired based on number of spectra (each of which
consists of a number of sweeps) are now based on duration (measure
for this many hours - each spectrum is acquired over X minutes).
Enabled the logging module.